Update payment mechanism to use withdrawals
- Move some presets to constants
- Add a queue of builder pending payments and a queue of builder pending
  withdrawals to the beacon state
- Add a fee recipient to the builder's bid
- Modify process epoch to either delete the builder's pending payment if
  the block was not voted enough, or queue a withdrawal if the block was
voted enough. Every epoch the pending payments from the previous epoch
are all processed (unless the payload was already included in which case
the payment was already processed) and the current epoch payments are
moved to the beginning of the queue.
- Modify the withdrawal handling by first processing all withdrawals
  from the builder's pending queue, before any other withdrawal. These
withdrawals are processed as follows. If the builder is slashed then
they are kept in the queue until the builder is withdrawable, in which
case the payments will be fullfilled with the remaining balance.
- Modify how to process the bid in the beacon block. Instead of
  deducting the builder immediately, we set a pending payment in the
queue, to be deducted later with the churn when processing the
withdrawal.
- Modify process attestations to keep track of those attestations that
  voted for the beacon block when no payload was included.